refactor: TeamResultResponse 가격관련 필드 Int -> Long으로 수정

+++ src/main/kotlin/team/mozu/dsm/adapter/in/team/dto/response/TeamResultResponse.kt
@@ -5,8 +5,8 @@ import java.util.UUID
 data class TeamResultResponse(
     val id: UUID,
     val teamName: String?,
-    val baseMoney: Int,
-    val totalMoney: Int,
+    val baseMoney: Long,
+    val totalMoney: Long,
     val invRound: Int,
     val valProfit: Int,
     val profitNum: Double,
